backend/requirements.txt
# ============================================================
# Beraxis AI Calling Platform — Python Dependencies
# ============================================================

# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Caching
redis[hiredis]==5.0.4
aioredis==2.0.1

# Celery (Background Tasks)
celery==5.3.6
celery[redis]==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7

# Rate Limiting
slowapi==0.1.9
limits==3.12.0

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1

# File Handling
python-magic==0.4.27
boto3==1.34.95
pillow==10.3.0

# Billing
stripe==9.5.0

# WebSockets
websockets==12.0

# Telephony — Asterisk ARI
requests==2.31.0

# AI — Speech to Text (Faster-Whisper)
# faster-whisper==1.0.3 (Comment out for lean local API testing)
numpy==1.26.4

# AI — LLM (Llama.cpp Python bindings)
# llama-cpp-python==0.2.76 (Comment out for lean local API testing)

# AI — TTS (Piper)
# piper-tts==1.2.0 (Fails on some Windows setups, comment out for local REST API testing)

# Audio Processing
soundfile==0.12.1
librosa==0.10.2
pydub==0.25.1

# Email
fastapi-mail==1.4.1
jinja2==3.1.4

# Monitoring & Logging
prometheus-fastapi-instrumentator==7.0.0
structlog==24.1.0
sentry-sdk[fastapi]==2.2.1

# Dev & Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
faker==24.14.0

# CORS & Middleware
starlette==0.37.2
